Almost evrything has changed for client side code
So to start off I added the lab 2 code you posted so everything that was required is now fully implemented. For assignment 1 I implemented a player controller script for movement on x and y axis and the cube is now a prefab. For NetworkMan Script I implemented a lot, sending the server position data, receiving other players position data and updating it in game. The colours no longer spam they are just set to 1 and that's it.  I set the player controler to specifically 1 player so others cant control eachother.